requirements.txt
fastapi
dateutil
mysql.connector
